+++ lib/domain/diary/post_diary_page.dart
@@ -4,7 +4,7 @@ import 'package:pilll/domain/diary/diary_state.codegen.dart';
 import 'package:pilll/domain/diary/post_diary_store.dart';
 import 'package:pilll/domain/diary/post_diary_store_provider_family.dart';
 import 'package:pilll/entity/diary.codegen.dart';
-import 'package:pilll/service/diary.dart';
+import 'package:pilll/database/diary.dart';
 import 'package:pilll/components/atoms/buttons.dart';
 import 'package:pilll/components/atoms/color.dart';
 import 'package:pilll/components/atoms/font.dart';
