+++ lib/domain/diary/post_diary_store.dart
@@ -1,6 +1,6 @@
 import 'package:pilll/domain/diary/diary_state.codegen.dart';
 import 'package:pilll/entity/diary.codegen.dart';
-import 'package:pilll/service/diary.dart';
+import 'package:pilll/database/diary.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
 class PostDiaryStore extends StateNotifier<DiaryState> {
